requirements.txt
Django

djangorestframework
djangorestframework-jwt>=1.8.0

django-extensions
django-filter
django-guardian
drfdocs
